pyschool/requirements.txt
anyio
asgiref
assertpy
attrs
bcrypt
certifi
cffi
charset-normalizer
click
datasets
dnspython
email-validator
fabric
fastapi
fastapi-rfc7807
h11
httptools
hypothesis
idna
iniconfig
itsdangerous
Jinja2
langchain
langchain-openai
langchain-ollama
langchainhub
langgraph
langserve[all]
MarkupSafe
orjson
packaging
pandas
passlib
pluggy
py
pycparser
pydantic
pyparsing
pyrsistent
pytest
python-dotenv
python-multipart
PyYAML
requests
six
sniffio
SQLAlchemy
sqlalchemy2-stubs
sqlmodel
starlette
streamlit
tomli
typer
typing_extensions
ujson
unittest2
urllib3
uvicorn
watchgod
websockets
xmlschema